CRUD Acess Implemnet

+++ Backend/V1/API/graphql/resolvers/admins.resolvers.js
@@ -15,6 +15,9 @@ const resolvers = {
       if (!context.user) throw new AuthenticationError("Authentication required");
       if (context.user.role !== 'super-admin') throw new ForbiddenError("Not authorized");
 
+      if (context.user.role === 'super-admin' && !context.user.status) {
+        throw new ForbiddenError("Master do not have access to this resource");
+      }
       return await adminsController.getAdminByID(id);
     }
   },
@@ -24,6 +27,9 @@ const resolvers = {
       if (!context.user) throw new AuthenticationError("Authentication required");
       if (context.user.role !== 'super-admin') throw new ForbiddenError("Only super-admins can create admins");
 
+      if (context.user.role === 'super-admin' && !context.user.status) {
+        throw new ForbiddenError("Master do not have access to this resource");
+      }
       const superAdminId = context.user.id;
       return await adminsController.RegisterAdmins(input, superAdminId);
     },
@@ -36,6 +42,9 @@ const resolvers = {
       if (!context.user) throw new AuthenticationError("Authentication required");
       if (context.user.role !== 'super-admin') throw new ForbiddenError("Not authorized");
 
+      if (context.user.role === 'super-admin' && !context.user.status) {
+        throw new ForbiddenError("Master do not have access to this resource");
+      }
       return await adminsController.updateAdmin(id, input);
     },
 
@@ -43,6 +52,10 @@ const resolvers = {
       if (!context.user) throw new AuthenticationError("Authentication required");
       if (context.user.role !== 'super-admin') throw new ForbiddenError("Not authorized");
 
+
+      if (context.user.role === 'super-admin' && !context.user.status) {
+        throw new ForbiddenError("Master do not have access to this resource");
+      }
       return await adminsController.deleteAdmin(id);
     }
   }
