+++ Backend/V1/API/graphql/resolvers/student.resolvers.js
@@ -7,7 +7,7 @@ const resolvers = {
   Query: {
     getAllStudents: async (_, __, context) => {
       if (!context.user) throw new AuthenticationError('Authentication required');
-      if (!['super-admin', 'admin', 'Teacher'].includes(context.user.role)) {
+      if (!['super-admin', 'admin', 'Teacher', 'parent'].includes(context.user.role)) {
         throw new ForbiddenError('Only super-admin, admin, and teacher can getAllStudents');
       }
       return await studentController.getAllStudents();
@@ -19,9 +19,16 @@ const resolvers = {
         throw new ForbiddenError('Only super-admin, admin, and teacher can getStudentById');
       }
 
+      if (context.user.role === 'super-admin' && !context.user.status) {
+        throw new ForbiddenError("Master do not have access to this resource");
+      }
       if (context.user.role === 'Teacher' && !context.user.status) {
         throw new ForbiddenError("You do not have access to this resource");
       }
+      if (context.user.role === 'parent' && !context.user.status) {
+        throw new ForbiddenError("parent do not have access to this resource");
+      }
+
       return await studentController.getStudentById(id);
     }
   },
@@ -29,12 +36,15 @@ const resolvers = {
   Mutation: {
     registerStudent: async (_, { input }, context) => {
       if (!context.user) throw new AuthenticationError('Authentication required');
-      if (!['super-admin', 'admin', 'Teacher'].includes(context.user.role)) {
+      if (!['super-admin', 'admin', 'Teacher',].includes(context.user.role)) {
         throw new ForbiddenError('Only super-admin, admin, and teacher can registerStudent');
       }
 
+      if (context.user.role === 'super-admin' && !context.user.status) {
+        throw new ForbiddenError("Master do not have access to this resource");
+      }
       if (context.user.role === 'Teacher' && !context.user.status) {
-        throw new ForbiddenError("You do not have access to this resource");
+        throw new ForbiddenError("Teacher do not have access to this resource");
       }
 
       const createdByID = context.user?.id || null;
@@ -49,9 +59,13 @@ const resolvers = {
         throw new ForbiddenError('Only super-admin, admin, and teacher can updateStudent');
       }
 
+      if (context.user.role === 'super-admin' && !context.user.status) {
+        throw new ForbiddenError("Master do not have access to this resource");
+      }
       if (context.user.role === 'Teacher' && !context.user.status) {
         throw new ForbiddenError("You do not have access to this resource");
       }
+
       return await studentController.updateStudent(id, input);
     },
 
@@ -61,8 +75,14 @@ const resolvers = {
         throw new ForbiddenError('Only super-admin, admin, and teacher can deleteStudent');
       }
 
+      if (context.user.role === 'super-admin' && !context.user.status) {
+        throw new ForbiddenError("Master do not have access to this resource");
+      }
       if (context.user.role === 'Teacher' && !context.user.status) {
         throw new ForbiddenError("You do not have access to this resource");
+      }
+      if (context.user.role === 'parent' && !context.user.status) {
+        throw new ForbiddenError("parent do not have access to this resource");
       }
       return await studentController.deleteStudent(id);
     }
